requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
Flask>=3.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
scikit-learn>=1.3.0
pandas>=2.0.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pgvector>=0.2.0
numpy>=1.24.0
